+++ src/Sql/Sql/Database/Cmdlet/NewAzureSqlDatabase.cs
@@ -197,16 +197,9 @@ namespace Microsoft.Azure.Commands.Sql.Database.Cmdlet
         /// <summary>
         /// Gets or sets the number of readonly replicas for the Azure Sql database
         /// </summary>
-        [CmdletParameterBreakingChange("ReadReplicaCount", ReplaceMentCmdletParameterName = "HighAvailabilityReplicaCount")]
         [Parameter(Mandatory = false,
             HelpMessage = "The number of readonly secondary replicas associated with the database to which readonly application intent connections may be routed. This property is only settable for Hyperscale edition databases.")]
-        public int ReadReplicaCount { get; set; }
-
-        /// <summary>
-        /// Gets or sets the number of readonly replicas for the Azure Sql database
-        /// </summary>
-        [Parameter(Mandatory = false,
-            HelpMessage = "The number of readonly secondary replicas associated with the database to which readonly application intent connections may be routed. This property is only settable for Hyperscale edition databases.")]
+        [Alias("ReadReplicaCount")]
         public int HighAvailabilityReplicaCount { get; set; }
 
         /// <summary>
@@ -305,8 +298,7 @@ namespace Microsoft.Azure.Commands.Sql.Database.Cmdlet
                 LicenseType = LicenseType, // note: default license type will be LicenseIncluded in SQL RP if not specified
                 AutoPauseDelayInMinutes = this.IsParameterBound(p => p.AutoPauseDelayInMinutes) ? AutoPauseDelayInMinutes : (int?)null,
                 MinimumCapacity = this.IsParameterBound(p => p.MinimumCapacity) ? MinimumCapacity : (double?)null,
-                HighAvailabilityReplicaCount = this.IsParameterBound(p => p.HighAvailabilityReplicaCount) ? HighAvailabilityReplicaCount : 
-                    this.IsParameterBound(p => p.ReadReplicaCount) ? ReadReplicaCount : (int?)null,
+                HighAvailabilityReplicaCount = this.IsParameterBound(p => p.HighAvailabilityReplicaCount) ? HighAvailabilityReplicaCount : (int?)null,
                 BackupStorageRedundancy = BackupStorageRedundancy,
                 SecondaryType = SecondaryType,
             };