+++ src/Sql/Sql/Database/Cmdlet/SetAzureSqlDatabase.cs
@@ -220,24 +220,13 @@ namespace Microsoft.Azure.Commands.Sql.Database.Cmdlet
         /// <summary>
         /// Gets or sets the number of read replicas for the database
         /// </summary>
-        [CmdletParameterBreakingChange("ReadReplicaCount", ReplaceMentCmdletParameterName = "HighAvailabilityReplicaCount")]
         [Parameter(Mandatory = false,
             HelpMessage = "The number of readonly secondary replicas associated with the database.  For Hyperscale edition only.",
             ParameterSetName = UpdateParameterSetName)]
         [Parameter(Mandatory = false,
             HelpMessage = "The number of readonly secondary replicas associated with the database.  For Hyperscale edition only.",
             ParameterSetName = VcoreDatabaseParameterSet)]
-        public int ReadReplicaCount { get; set; }
-
-        /// <summary>
-        /// Gets or sets the number of read replicas for the database
-        /// </summary>
-        [Parameter(Mandatory = false,
-            HelpMessage = "The number of readonly secondary replicas associated with the database.  For Hyperscale edition only.",
-            ParameterSetName = UpdateParameterSetName)]
-        [Parameter(Mandatory = false,
-            HelpMessage = "The number of readonly secondary replicas associated with the database.  For Hyperscale edition only.",
-            ParameterSetName = VcoreDatabaseParameterSet)]
+        [Alias("ReadReplicaCount")]
         public int HighAvailabilityReplicaCount { get; set; }
 
         /// <summary>
@@ -306,8 +295,7 @@ namespace Microsoft.Azure.Commands.Sql.Database.Cmdlet
                 LicenseType = LicenseType ?? model.FirstOrDefault().LicenseType, // set to original license type
                 AutoPauseDelayInMinutes = this.IsParameterBound(p => p.AutoPauseDelayInMinutes) ? AutoPauseDelayInMinutes : (int?)null,
                 MinimumCapacity = this.IsParameterBound(p => p.MinimumCapacity) ? MinimumCapacity : (double?)null,
-                HighAvailabilityReplicaCount = this.IsParameterBound(p => p.HighAvailabilityReplicaCount) ? HighAvailabilityReplicaCount : 
-                    this.IsParameterBound(p => p.ReadReplicaCount) ? ReadReplicaCount : (int?)null,
+                HighAvailabilityReplicaCount = this.IsParameterBound(p => p.HighAvailabilityReplicaCount) ? HighAvailabilityReplicaCount : (int?)null,
                 BackupStorageRedundancy = BackupStorageRedundancy,
                 SecondaryType = SecondaryType,
             };